requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
waitress==2.1.2
openai
